requirements/base.txt
django
Pillow
djangorestframework
django-environ
django-bootstrap4
django-crispy-forms
psycopg2